the interface that will provide basic player data such as statistics and more
